[ADQL] Restore `REGION(...)` & Add `CAST(...)` & JavaCC-7.0.10 in Gradle
This time `REGION(...)` can only accept string literal as single parameter and
only the DALI and the STC/s serializations are allowed. It is however possible
to accept any string expression and/or allow more serializations ; to do that,
`ADQLParser.allowExtendedRegionParam(true)`.

In the current state `CAST(...)` supports only the standard datatypes and no
geometrical one. In a further version, it may be possible to specify custom
datatypes and how to translate/cast them.

The JavaCC's grammar files are now compiled with JavaCC-7.0.10 (released in
Nov. 2020). Generated files like ParseException, TokenManager, etc have been
upgraded too, while keeping the customization required by the VOLLT/ADQLLib.
The generation of JavaCC files is now possible with Gradle through the tasks
`build` and `javacc`.

Also fixed:

- Custom type creation while parsing an unknown type in the UDF definition
  (FunctionDef.parse) : the length attribute was included in the type's name
  instead of being apart (ex: `custom(10)` was interpreted as the custom type
  named `custom(10)` instead of `custom` with the length '10').

- Fix error message when a MathFunction has more parameter as expected.